add a few internal helper methods.
remove_node_dirty allows us to remove a node without cleaning up after ourselves.
this should be faster when dropping multiple generations of nodes when we don't care what happens to the nodes.  It needs to be used carefully however, as it can be dangerous if not used properly.
add drop_children_recursive that does what is described above.